Add SARIF defaultConfiguration level support
