+++ src/main/java/de/tum/cit/aet/artemis/programming/service/localci/scaparser/format/sarif/Result.java
@@ -24,15 +24,14 @@ import com.fasterxml.jackson.annotation.JsonValue;
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record Result(String ruleId, Integer ruleIndex, ReportingDescriptorReference rule,
-        de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.Result.Kind kind,
-        de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.Result.Level level, Message message, List<Location> locations) {
+        de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.Result.Kind kind, Level level, Message message, List<Location> locations) {
 
     public Result(String ruleId, Integer ruleIndex, ReportingDescriptorReference rule, Kind kind, Level level, Message message, List<Location> locations) {
         this.ruleId = ruleId;
         this.ruleIndex = Objects.requireNonNullElse(ruleIndex, -1);
         this.rule = rule;
         this.kind = Objects.requireNonNullElse(kind, Kind.FAIL);
-        this.level = Objects.requireNonNullElse(level, Level.WARNING);
+        this.level = level;
         this.message = message;
         this.locations = locations;
     }
@@ -129,57 +128,6 @@ public record Result(String ruleId, Integer ruleIndex, ReportingDescriptorRefere
                 return constant;
             }
         }
-    }
-
-    /**
-     * A value specifying the severity level of the result.
-     */
-    public enum Level {
-
-        NONE("none"), NOTE("note"), WARNING("warning"), ERROR("error");
-
-        private final String value;
-
-        private static final Map<String, Level> CONSTANTS = new HashMap<>();
-
-        static {
-            for (Level c : values()) {
-                CONSTANTS.put(c.value, c);
-            }
-        }
-
-        Level(String value) {
-            this.value = value;
-        }
-
-        @Override
-        public String toString() {
-            return this.value;
-        }
-
-        @JsonValue
-        public String value() {
-            return this.value;
-        }
-
-        /**
-         * Creates a {@link Level} instance from a given string value.
-         *
-         * @param value the string representation of the {@link Level}
-         * @return the matching {@link Level} instance
-         * @throws IllegalArgumentException if the provided value does not correspond to any defined {@link Level}
-         */
-        @JsonCreator
-        public static Level fromValue(String value) {
-            Level constant = CONSTANTS.get(value);
-            if (constant == null) {
-                throw new IllegalArgumentException(value);
-            }
-            else {
-                return constant;
-            }
-        }
-
     }
 
 }
