+++ src/main/java/de/tum/cit/aet/artemis/programming/service/localci/scaparser/strategy/sarif/SarifParser.java
@@ -19,10 +19,12 @@ import de.tum.cit.aet.artemis.programming.dto.StaticCodeAnalysisIssue;
 import de.tum.cit.aet.artemis.programming.dto.StaticCodeAnalysisReportDTO;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.ArtifactLocation;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.GlobalMessageStrings;
+import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.Level;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.Location;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.MessageStrings;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.PhysicalLocation;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.Region;
+import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.ReportingConfiguration;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.ReportingDescriptor;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.ReportingDescriptorReference;
 import de.tum.cit.aet.artemis.programming.service.localci.scaparser.format.sarif.Result;
@@ -120,7 +122,7 @@ public class SarifParser implements ParserStrategy {
         // Fallback to the rule identifier for the category
         String category = rule.map(ruleCategorizer::categorizeRule).orElse(ruleId);
 
-        Result.Level level = result.getOptionalLevel().orElse(Result.Level.WARNING);
+        Level level = result.getOptionalLevel().orElseGet(() -> getDefaultLevel(rule));
 
         String message = findMessage(result, driver, rule);
 
@@ -182,4 +184,8 @@ public class SarifParser implements ParserStrategy {
         return Optional.of(baseRuleId);
     }
 
+    @SuppressWarnings("OptionalUsedAsFieldOrParameterType")
+    private static Level getDefaultLevel(Optional<ReportingDescriptor> rule) {
+        return rule.flatMap(ReportingDescriptor::getOptionalDefaultConfiguration).flatMap(ReportingConfiguration::getOptionalLevel).orElse(Level.WARNING);
+    }
 }
